python/FXModeling/requirements.txt
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
arch>=6.0.0
torch>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.28.0